SG3_Devices/connect/gateway/requirements.txt
anyio==4.12.1
blinker==1.9.0
CacheControl==0.14.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.5
firebase_admin==7.1.0
Flask==3.1.3
google-api-core==2.30.0
google-auth==2.48.0
google-cloud-core==2.5.0
google-cloud-firestore==2.23.0
google-cloud-storage==3.9.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.78.1
grpcio-status==1.78.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
msgpack==1.1.2
proto-plus==1.27.1
protobuf==6.33.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.11.0
pyserial==3.5
python-dotenv==1.2.1
requests==2.32.5
rsa==4.9.1
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.6
pyserial
